feat(mcp): extract registry normalization into registry-normalize-lib
Move MCP platform alias resolution and pagination normalization helpers
into a focused lib module with targeted tests.
